refactor: extract the OSV client and add per-source scan status
Extract the OSV API client out of dependency-analysis into a shared
server/utils/osv.ts, and additively extend the vulnerabilities API
response with per-source metadata:

- VulnerabilityTreeResult gains sourceStatus ({ osv: 'ok' | 'partial'
  | 'failed' }) so consumers can distinguish "no known vulnerabilities"
  from "could not check"
- VulnerabilitySummary entries are tagged with the sources that
  reported them
- cache keys bumped (route vulnerabilities:v1 -> v2, function v2 -> v3)
  for the new response shape
- results degraded by a transient source failure are only served from
  the dependency-analysis cache for five minutes instead of an hour,
  and the route response cache now matches its 300s ISR rule, so an
  OSV blip doesn't strip findings from caches long after recovery

The new metadata is additive; consuming it in the UI is a separate
change.
